requirements.txt
pandas
numpy
seaborn
scikit-learn
catboost
xgboost
dill
flask
tensorflow
dvc
notebook
python-box==6.0.2
tqdm
pyYAML
scipy
ensure==1.0.2
boto3
s3fs
-e .